refactor: migrate product and payment services to new API structure
