--- objects/product.py
@@ -1,24 +1,0 @@
-# Backward-compatibility imports
-from app.models.stock import                Stock
-from app.models.category import             Category
-from app.models.product import              Product
-from app.models.supplier import             Supplier
-from app.models.brand import                Brand
-from app.models.description import          Description
-from app.models.productImage import         ProductImage
-from app.models.categoryProductLink import  CategoryProductLink
-from app.models.productReview import        ProductReview
-
-
-
-__all__ = [
-    "CategoryProductLink",
-    "Category",
-    "Supplier",
-    "Brand",
-    "Description",
-    "Stock",
-    "Product",
-    "ProductReview",
-    "ProductImage",
-]
