--- routers/products.py
@@ -1,3 +1,0 @@
-from app.features.products.router import router
-
-__all__ = ["router"]
